refactor: move content queries to api
